Increase timeout for docker compose down
The default timeout is 10s when waiting for services to stop via `docker compose down`. Services often took longer than this, which led to us returning from the command early, and then prematurely deleting the compose YAML file. The command still finished in the background, but now errored out with a "compose file at path not found" error.